Add Vortex, Pow, Subsonic packages via Xcode
- Vortex 1.0.4 (particle effects)
- Pow 1.0.5 (transitions)
- Subsonic 0.2.0 (audio)

Packages added correctly through Xcode UI.

+++ RenaissanceArchitectAcademy/Views/BloomEffectView.swift
@@ -101,7 +101,7 @@ struct SparkleParticle: View {
                 x: cos(angle.radians) * distance,
                 y: sin(angle.radians) * distance
             )
-            .opacity(1 - progress * 0.8)
+            .opacity(Double(1.0 - progress * 0.8))
             .blur(radius: progress * 2)
     }
 }
